moved getContainerId into demo script
