--- examples/apps/cra-demo/src/getContainerId.ts
@@ -1,14 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License.
- */
-/* eslint-disable no-restricted-globals */
-export const getContainerId = (): { containerId: string; isNew: boolean } => {
-    let isNew = false;
-    if (location.hash.length === 0) {
-        isNew = true;
-        location.hash = Date.now().toString();
-    }
-    const containerId = location.hash.substring(1);
-    return { containerId, isNew };
-};
